Move custom and image methods
